correcciones, se divide codigo python en dos archivos conexion y principal
